-implemented db -minor fixes

+++ src/main/kotlin/it/pureorigins/fancyparticles/particles/shapes/ParticleShape.kt
@@ -5,6 +5,6 @@ import net.minecraft.util.math.Vec3d
 interface ParticleShape {
     fun getPosition(iteration: Int): Vec3d
     fun getOffset(iteration: Int): Vec3d
-    fun getSpeed(iteration: Int): Int
+    fun getSpeed(iteration: Int): Double
     fun getCount(iteration: Int): Int
 }
